# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recipes/from-video:
    post:
      operationId: analyzeRecipeFromVideo
      tags: [video-import]
      summary: Analyze a cooking video into a recipe draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoRecipeImportRequest"
      responses:
        "202":
          description: Video analysis started; poll the job until complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoRecipeImportJobResponse"
        "400":
          description: Invalid video source
        "401":
          description: Authentication required
        "413":
          description: Video is too large
        "422":
          description: Video could not be analyzed
  /recipes/from-video/jobs/{jobId}:
    get:
      operationId: getVideoRecipeImportJob
      tags: [video-import]
      summary: Get the status of a cooking video recipe import
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed AI-generated recipe draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoRecipeImportResponse"
        "202":
          description: Video analysis is still processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoRecipeImportJobResponse"
        "401":
          description: Authentication required
        "404":
          description: Video analysis job not found
        "422":
          description: Video could not be analyzed
...
components:
  schemas:
...
    Ingredient:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
        aisle:
          type: string
        shoppingUrl:
          type: string
        price:
          type: number
          minimum: 0
      required: [name, quantity, aisle, shoppingUrl, price]
...
    RecipeInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        imageUrl:
          type: string
        category:
          type: string
          minLength: 1
        dietary:
          type: array
          items:
            type: string
        prepMinutes:
          type: number
          minimum: 0
        cookMinutes:
          type: number
          minimum: 0
        servings:
          type: number
          minimum: 1
        calories:
          type: number
          minimum: 0
        protein:
          type: number
          minimum: 0
        carbs:
          type: number
          minimum: 0
        fat:
          type: number
          minimum: 0
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        instructions:
          type: array
          items:
            type: string
        author:
          type: string
          minLength: 1
      required:
        - title
        - description
        - imageUrl
        - category
        - dietary
        - prepMinutes
        - cookMinutes
        - servings
        - calories
        - protein
        - carbs
        - fat
        - ingredients
        - instructions
        - author
...
    VideoRecipeImportRequest:
      type: object
      required: [sourceType, source]
      properties:
        sourceType:
          type: string
          enum: [url, upload]
        source:
          type: string
          minLength: 1
        mimeType:
          type: string
    VideoRecipeImportResponse:
      type: object
      required: [recipe, sourceLabel]
      properties:
        recipe:
          $ref: "#/components/schemas/RecipeInput"
        sourceLabel:
          type: string
    VideoRecipeImportJobResponse:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [processing, complete, failed]
        recipe:
          $ref: "#/components/schemas/RecipeInput"
        sourceLabel:
          type: string
        error:
          type: string