# lib/api-zod/src/generated/types/videoRecipeImportJobResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeInput } from './recipeInput';
import type { VideoRecipeImportJobResponseStatus } from './videoRecipeImportJobResponseStatus';

export interface VideoRecipeImportJobResponse {
  jobId: string;
  status: VideoRecipeImportJobResponseStatus;
  recipe?: RecipeInput;
  sourceLabel?: string;
  error?: string;
}