# lib/api-zod/src/generated/types/videoRecipeImportRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoRecipeImportRequestSourceType } from './videoRecipeImportRequestSourceType';

export interface VideoRecipeImportRequest {
  sourceType: VideoRecipeImportRequestSourceType;
  /** @minLength 1 */
  source: string;
  mimeType?: string;
}