# lib/api-zod/src/generated/types/videoRecipeImportResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeInput } from './recipeInput';

export interface VideoRecipeImportResponse {
  recipe: RecipeInput;
  sourceLabel: string;
}